RAG_1/requirements.txt
# Core RAG Dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# Vector Store
chromadb==0.4.22
sentence-transformers==2.3.1

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
python-multipart==0.0.6

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM Integration
openai==1.10.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0

# External Enrichment (Stretch)
wikipedia==1.4.0
duckduckgo-search==4.1.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

